@@ -21,23 +21,20 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/cache": "^11.11.0",
     "@emotion/react": "^11.10.6",
     "@emotion/server": "^11.10.0",
     "@emotion/styled": "^11.10.6",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/lab": "^5.0.0-alpha.126",
-    "@mui/material": "^5.12.0",
     "@tinacms/cli": "^1.5.6",
     "@types/node": "18.15.11",
     "@types/react": "18.0.35",
     "@types/react-dom": "18.0.11",
     "date-fns": "^2.29.3",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "next": "13.3.0",
+    "next": "^13.4.13",
     "next-seo": "^6.0.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tinacms": "^1.4.6",
     "voca": "^1.4.0"
   },
@@ -48,8 +45,9 @@
     "@types/react-dom": "18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.58.0",
     "@typescript-eslint/parser": "^5.58.0",
+    "autoprefixer": "^10.4.14",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
+    "eslint-config-next": "^13.4.13",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsdoc": "^41.1.1",
@@ -59,9 +57,11 @@
     "husky": "^8.0.3",
     "lint-staged": "^13.2.1",
     "next-sitemap": "^4.0.7",
+    "postcss": "^8.4.27",
     "prettier": "^2.8.7",
     "pretty-quick": "^3.1.3",
     "rimraf": "^5.0.0",
+    "tailwindcss": "^3.3.3",
     "typescript": "5.0.4"
   }
 }
