@@ -1,16 +1,10 @@
+"use client";
 import * as React from "react";
 
-import Styles from "./header.module.css";
-
 import Image from "next/image";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import NextLink from "next/link";
 
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
-import Toolbar from "@mui/material/Toolbar";
-
-import Link from "@/components/link/link";
 import PillButton from "@/components/pill-button/pill-button";
 
 interface NavItems {
@@ -38,49 +32,47 @@ const navItems: NavItems[] = [
 ];
 
 function Header() {
-  const router = useRouter();
+  const pathname = usePathname();
 
   const navLinks = (navigationItems: NavItems[]) => {
     return navigationItems.map((item: NavItems) => {
       return item.linkName === "Let's Talk" ? (
-        <Box key={item.linkName} className={Styles.btn_wrap}>
-          <PillButton title={item.linkName} className={Styles.btn} href={"/contact"} />
-        </Box>
+        <div key={item.linkName} className={"py-0 px-[0.5rem] sm:px-[1rem]"}>
+          <PillButton
+            title={item.linkName}
+            className={"py-[0.375rem] px-[0.75rem]"}
+            href={"/contact"}
+          />
+        </div>
       ) : (
-        <Box key={item.linkName} className={Styles.nav_wrap}>
-          <Link
+        <div key={item.linkName} className={"py-0 px-[0.5rem] sm:px-[1rem]"}>
+          <NextLink
             href={item.linkHref}
-            underline="none"
-            color={router.pathname === item.linkHref ? "primary.main" : "info.main"}
-            sx={{
-              ":hover": {
-                color: "primary.main",
-              },
-            }}
-            className={Styles.link}
+            className={`no-underline ${
+              pathname === item.linkHref
+                ? "text-primary-main"
+                : "text-info-main"
+            } hover:text-primary-main text-base`}
           >
             {item.linkName}
-          </Link>
-        </Box>
+          </NextLink>
+        </div>
       );
     });
   };
 
   return (
-    <Box className={Styles.container}>
-      <AppBar position="relative" component="nav" className={Styles.app_bar}>
-        <Toolbar>
-          {/* <IconButton
-            color="info"
-            aria-label="open drawer"
-            edge="start"
-            onClick={handleDrawerToggle}
-            sx={{ mr: 2, display: { sm: "none" } }}
-          >
-            <MenuIcon />
-          </IconButton> */}
-          <Box component="div" className={Styles.logo_wrap}>
-            <NextLink href={"/"} className={Styles.logo_link}>
+    <div className={"flex"}>
+      <nav
+        className={`relative flex flex-col w-full box-border flex-shrink-0 relative text-white bg-transparent justify-center h-[3.25rem] sm:h-[7.25rem] sm:py-[0] sm:px-[2.4rem]`}
+      >
+        <div
+          className={
+            "relative flex items-center pl-4 pr-4 min-h-[2.875rem] sm:min-h-[4rem] sm:pl-6 sm:pr-6"
+          }
+        >
+          <div className={"hidden sm:block sm:flex-grow sm:items-center"}>
+            <NextLink href={"/"} className={"flex"}>
               <Image
                 src="/images/logos/octalogic.svg"
                 alt="Octalogic logo"
@@ -88,31 +80,13 @@ function Header() {
                 height={60}
               />
             </NextLink>
-          </Box>
-          <Box className={Styles.links_wrap}>{navLinks(navItems)}</Box>
-        </Toolbar>
-      </AppBar>
-      {/* <Box component="nav">
-        <Drawer
-          container={container}
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{
-            keepMounted: true, // Better open performance on mobile.
-          }}
-          sx={{
-            display: { xs: "block", sm: "none" },
-            "& .MuiDrawer-paper": {
-              boxSizing: "border-box",
-              width: drawerWidth,
-            },
-          }}
-        >
-          {drawer}
-        </Drawer>
-      </Box> */}
-    </Box>
+          </div>
+          <div className={"hidden sm:flex sm:flex-row sm:items-center"}>
+            {navLinks(navItems)}
+          </div>
+        </div>
+      </nav>
+    </div>
   );
 }
 
