@@ -1,52 +1,50 @@
 import * as React from "react";
 import clsx from "clsx";
-import { useRouter } from "next/router";
+import { useRouter, usePathname } from "next/navigation";
 import NextLink, { LinkProps as NextLinkProps } from "next/link";
-import MuiLink, { LinkProps as MuiLinkProps } from "@mui/material/Link";
-import { styled } from "@mui/material/styles";
-// import styles from './link.module.css';
-
-// Add support for the sx prop for consistency with the other branches.
-const Anchor = styled("a")({});
 
 interface NextLinkComposedProps
   extends Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, "href">,
-    Omit<NextLinkProps, "href" | "as" | "onClick" | "onMouseEnter" | "onTouchStart"> {
+    Omit<
+      NextLinkProps,
+      "href" | "as" | "onClick" | "onMouseEnter" | "onTouchStart"
+    > {
   to: NextLinkProps["href"];
   linkAs?: NextLinkProps["as"];
 }
 
-export const NextLinkComposed = React.forwardRef<HTMLAnchorElement, NextLinkComposedProps>(
-  function NextLinkComposed(props, ref) {
-    const {
-      to,
-      linkAs,
-      replace,
-      scroll,
-      shallow,
-      prefetch,
-      legacyBehavior = true,
-      locale,
-      ...other
-    } = props;
+export const NextLinkComposed = React.forwardRef<
+  HTMLAnchorElement,
+  NextLinkComposedProps
+>(function NextLinkComposed(props, ref) {
+  const {
+    to,
+    linkAs,
+    replace,
+    scroll,
+    shallow,
+    prefetch,
+    legacyBehavior = true,
+    locale,
+    ...other
+  } = props;
 
-    return (
-      <NextLink
-        href={to}
-        prefetch={prefetch}
-        as={linkAs}
-        replace={replace}
-        scroll={scroll}
-        shallow={shallow}
-        passHref
-        locale={locale}
-        legacyBehavior={legacyBehavior}
-      >
-        <Anchor ref={ref} {...other} />
-      </NextLink>
-    );
-  },
-);
+  return (
+    <NextLink
+      href={to}
+      prefetch={prefetch}
+      as={linkAs}
+      replace={replace}
+      scroll={scroll}
+      shallow={shallow}
+      passHref
+      locale={locale}
+      legacyBehavior={legacyBehavior}
+    >
+      <a ref={ref} {...other} />
+    </NextLink>
+  );
+});
 
 export type LinkProps = {
   activeClassName?: string;
@@ -54,12 +52,14 @@ export type LinkProps = {
   href: NextLinkProps["href"];
   linkAs?: NextLinkProps["as"]; // Useful when the as prop is shallow by styled().
   noLinkStyle?: boolean;
-} & Omit<NextLinkComposedProps, "to" | "linkAs" | "href"> &
-  Omit<MuiLinkProps, "href">;
+} & Omit<NextLinkComposedProps, "to" | "linkAs" | "href">;
 
 // A styled version of the Next.js Link component:
 // https://nextjs.org/docs/api-reference/next/link
-const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(function Link(props, ref) {
+const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(function Link(
+  props,
+  ref
+) {
   const {
     activeClassName = "active",
     as,
@@ -78,20 +78,18 @@ const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(function Link(props,
   } = props;
 
   const router = useRouter();
+  const currentPathname = usePathname();
   const pathname = typeof href === "string" ? href : href.pathname;
   const className = clsx(classNameProps, {
-    [activeClassName]: router.pathname === pathname && activeClassName,
+    [activeClassName]: currentPathname === pathname && activeClassName,
   });
 
   const isExternal =
-    typeof href === "string" && (href.indexOf("http") === 0 || href.indexOf("mailto:") === 0);
+    typeof href === "string" &&
+    (href.indexOf("http") === 0 || href.indexOf("mailto:") === 0);
 
   if (isExternal) {
-    if (noLinkStyle) {
-      return <Anchor className={className} href={href} ref={ref} {...other} />;
-    }
-
-    return <MuiLink className={className} href={href} ref={ref} {...other} />;
+    return <a className={className} href={href} ref={ref} {...other} />;
   }
 
   const linkAs = linkAsProp || as;
@@ -106,13 +104,8 @@ const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(function Link(props,
     locale,
   };
 
-  if (noLinkStyle) {
-    return <NextLinkComposed className={className} ref={ref} {...nextjsProps} {...other} />;
-  }
-
   return (
-    <MuiLink
-      component={NextLinkComposed}
+    <NextLinkComposed
       className={className}
       ref={ref}
       {...nextjsProps}
