@@ -1,7 +1,4 @@
-import Styles from "./footer.module.css";
-
-import { Typography } from "@mui/material";
-import Box from "@mui/material/Box";
+"use client";
 
 import Image from "next/image";
 import NextLink from "next/link";
@@ -14,32 +11,40 @@ import { ISocial } from "@/interfaces";
 
 const FooterLink = ({ name, href }: { name: string; href: string }) => {
   return (
-    <Box className={Styles.container}>
+    <div
+      className={
+        "normal-case text-center p-1 flex-[0_0_33.333333%] max-w-[33.333333%] sm:flex-[0_0_16.666667%] sm: max-w-[16.666667%]"
+      }
+    >
       <Link
         href={href}
-        underline="none"
         color="info.main"
-        className={Styles.link}
+        className={"text-base opacity-70 hover:opacity-100"}
+        style={{ transition: " opacity 0.2s" }}
       >
         {name}
       </Link>
-    </Box>
+    </div>
   );
 };
 
 const SocialLinks = ({ socials }: { socials: ISocial[] }) => {
   const elementRows = socials.map((social) => (
-    <Box key={social.name} className={Styles.social_wrap}>
+    <div
+      key={social.name}
+      className={"opacity-70 hover:opacity-100"}
+      style={{ transition: " opacity 0.2s" }}
+    >
       <a href={social.link} target="_blank" rel="noreferrer">
         <Image
           src={social.iconUrl}
           alt={`${social.name} logo`}
           width={16}
           height={16}
-          className={Styles.social_img}
+          className={"cursor-pointer"}
         />
       </a>
-    </Box>
+    </div>
   ));
 
   return <>{elementRows}</>;
@@ -49,42 +54,42 @@ export function Footer() {
   const date: Date = new Date();
 
   return (
-    <Box className={Styles.footer_container}>
-      <Box className={Styles.logo_wrap}>
+    <div className={"w-full h-[22.375rem] flex flex-col items-center"}>
+      <div className={"mt-12 flex"}>
         <NextLink href={"/"}>
           <Image
             src="/images/logos/octalogic.svg"
             alt="Octalogic logo"
             width={60}
             height={60}
-            className={Styles.logo_img}
+            className={"cursor-pointer"}
           />
         </NextLink>
-      </Box>
-      <Box className={Styles.links_container}>
-        <Box className={Styles.links_wrap}>
+      </div>
+      <div className={"w-full flex flex-col items-center my-12"}>
+        <div className={"w-full flex justify-center mb-[0.187rem]"}>
           <FooterLink name={"Contact"} href={"contact"} />
           <FooterLink name={"Home"} href={"/"} />
           <FooterLink name={"Remote Resources"} href={"remote-resources"} />
-        </Box>
-        <Box className={Styles.links_wrap}>
+        </div>
+        <div className={"w-full flex justify-center mb-[0.187rem]"}>
           <FooterLink name={"Privacy"} href={"privacy-policy"} />
           <FooterLink name={"About"} href={"about"} />
           <FooterLink name={"Web Dev"} href={"web-development"} />
-        </Box>
-        <Box className={Styles.links_wrap}>
+        </div>
+        <div className={"w-full flex justify-center mb-[0.187rem]"}>
           <FooterLink name={"Terms of Service"} href={"terms-of-service"} />
           <FooterLink name={"Services"} href={"services"} />
           <FooterLink name={"Mobile Dev"} href={"mobile-development"} />
-        </Box>
-      </Box>
-      <Box className={Styles.socials_container}>
+        </div>
+      </div>
+      <div className={"flex items-center mb-4 gap-4"}>
         <SocialLinks socials={Socials} />
-      </Box>
-      <Typography component="div" className={Styles.copyright}>
+      </div>
+      <div className={"text-[0.8rem]/[0.8rem] text-center pb-2"}>
         © 2017 - {date.getFullYear()}, Octalogic Tech LLP. All rights reserved
-      </Typography>
-    </Box>
+      </div>
+    </div>
   );
 }
 
